pyproject.toml
[tool.poetry]
name = "grpc-prometheus-metrics"
version = "0.0.4"
description = "Python gRPC Prometheus Interceptors, py-grpc-prometheus fork"
authors = ["Sergio Soldatov <soldatovsr97@gmail.com>", "Lin Chen <linchen04@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"
packages = [{include = "grpc_prometheus_metrics"}]

[tool.poetry.dependencies]
python = "^3.10"
setuptools = ">=39.0.1"
grpcio = ">=1.10.0"
prometheus-client = ">=0.3.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.0"
pre-commit = "^3.1.1"
pytest = "^7.2.2"
requests = "^2.28.2"
coverage = "^7.2.1"
grpcio-tools = "^1.48.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
